% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.ref.R
\name{color.ref}
\alias{color.ref}
\title{Display a Color Reference Palette}
\usage{
color.ref()
}
\value{
A plot displaying the color palette.
}
\description{
This function generates a plot displaying a predefined color palette with
color codes for easy reference. The palette includes shades of Red, Orange,
Yellow, Green, Blue, Purple, and Grey.
}
\examples{
color.ref()
}
