% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Dark.Orange}
\alias{Dark.Orange}
\title{Dark Orange color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Dark.Orange
}
\description{
Dark Orange color
}
\keyword{datasets}
