% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Dark.Purple}
\alias{Dark.Purple}
\title{Dark Purple color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Dark.Purple
}
\description{
Dark Purple color
}
\keyword{datasets}
