% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Dark.Yellow}
\alias{Dark.Yellow}
\title{Dark Yellow color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Dark.Yellow
}
\description{
Dark Yellow color
}
\keyword{datasets}
