% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Deep.Orange}
\alias{Deep.Orange}
\title{Deep Orange color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Deep.Orange
}
\description{
Deep Orange color
}
\keyword{datasets}
