% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Deep.Purple}
\alias{Deep.Purple}
\title{Deep Purple color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Deep.Purple
}
\description{
Deep Purple color
}
\keyword{datasets}
