% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Deep.Yellow}
\alias{Deep.Yellow}
\title{Deep Yellow color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Deep.Yellow
}
\description{
Deep Yellow color
}
\keyword{datasets}
