% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Light.Green}
\alias{Light.Green}
\title{Light Green color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Light.Green
}
\description{
Light Green color
}
\keyword{datasets}
