% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Light.Orange}
\alias{Light.Orange}
\title{Light Orange color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Light.Orange
}
\description{
Light Orange color
}
\keyword{datasets}
