% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Light.Purple}
\alias{Light.Purple}
\title{Light Purple color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Light.Purple
}
\description{
Light Purple color
}
\keyword{datasets}
