% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Light.Yellow}
\alias{Light.Yellow}
\title{Light Yellow color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Light.Yellow
}
\description{
Light Yellow color
}
\keyword{datasets}
