% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Pale.Orange}
\alias{Pale.Orange}
\title{Pale Orange color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Pale.Orange
}
\description{
Pale Orange color
}
\keyword{datasets}
