% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Pale.Purple}
\alias{Pale.Purple}
\title{Pale Purple color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Pale.Purple
}
\description{
Pale Purple color
}
\keyword{datasets}
