% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{Pale.Yellow}
\alias{Pale.Yellow}
\title{Pale Yellow color}
\format{
An object of class \code{character} of length 1.
}
\usage{
Pale.Yellow
}
\description{
Pale Yellow color
}
\keyword{datasets}
