% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge.regression.R
\name{print.ridge.model}
\alias{print.ridge.model}
\title{Print Method for Ridge Model Objects}
\usage{
\method{print}{ridge.model}(x, ...)
}
\arguments{
\item{x}{A `ridge.model` object}

\item{...}{Additional arguments (not used)}
}
\description{
Prints a summary of the ridge model fit.
}
