% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snazzieR.theme.R
\name{snazzieR.theme}
\alias{snazzieR.theme}
\title{A Custom ggplot2 Theme for Publication-Ready Plots}
\usage{
snazzieR.theme()
}
\value{
A ggplot2 theme object.
}
\description{
This theme provides a clean, polished look for ggplot2 plots, with a focus on
readability and aesthetics. It includes a custom color palette and formatting
for titles, axes, and legends.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  snazzieR.theme()
}
